messages: usage rendering and its tests in their own submodules; disclosure keeps caching provider-dependent
encode_messages.rs and encode_messages/tests.rs crossed the 999-line budget;
the usage object moves to encode_messages/usage.rs and its tests to
tests_usage.rs. The cache_control disclosure reads
->not_forwarded(provider_decides_caching; cache reads reported in
usage.cache_read_input_tokens): a generic OpenAI-compatible endpoint may never
cache, so the suffix names where any reads show up without claiming caching
is on (Greptile review).
