+++ exp/runtime/gateway/native/src/encode_messages.rs
@@ -87,56 +87,6 @@ pub(super) fn stop_sequence_value(terminal: &Event) -> Value {
     }
 }
 
-/// Anthropic's usage object for every Messages frame that carries one:
-/// `message_start.message.usage`, `message_delta.usage`, and the
-/// non-streamed body's `usage`. All four token legs are always present,
-/// `0` when the provider reported none, because Anthropic clients (the
-/// official SDK accumulators, Claude Code's context meter) read the cache
-/// legs by key and treat an absent key as "not Anthropic's shape".
-///
-/// Mapping from the normalized `Usage` (whose `input_tokens` is the FOLDED
-/// total the ledger bills: uncached + cache reads + cache writes):
-///
-/// | Anthropic field                | source                                              |
-/// |--------------------------------|-----------------------------------------------------|
-/// | `input_tokens`                 | `input_tokens - cached_input_tokens - cache_creation_input_tokens` (uncached input, saturating) |
-/// | `cache_creation_input_tokens`  | `cache_creation_input_tokens`, else `0` (Anthropic-wire rungs only) |
-/// | `cache_read_input_tokens`      | `cached_input_tokens`, else `0` (Anthropic `cache_read_input_tokens`, OpenAI-wire `prompt_tokens_details.cached_tokens` / `input_tokens_details.cached_tokens`, Gemini `cachedContentTokenCount`, Bedrock `cacheReadInputTokens`) |
-/// | `output_tokens`                | `output_tokens` (reasoning folded in where the provider bills it additively) |
-///
-/// Unknown usage (no provider report) renders every leg as `0`.
-pub(super) fn messages_usage(usage: Option<&Usage>) -> Value {
-    let usage = match usage {
-        Some(usage) if usage.has_token_counts() => usage,
-        _ => return usage_object(0, 0, 0, 0),
-    };
-    let cached = usage.cached_input_tokens.unwrap_or(0);
-    let creation = usage.cache_creation_input_tokens.unwrap_or(0);
-    // Both cache legs come back out of the folded ledger total so callers
-    // see the provider's own shape: input_tokens excludes cached reads and
-    // cache writes, each reported on its own leg.
-    usage_object(
-        usage
-            .input_tokens
-            .unwrap_or(0)
-            .saturating_sub(cached)
-            .saturating_sub(creation),
-        creation,
-        cached,
-        usage.output_tokens.unwrap_or(0),
-    )
-}
-
-/// The four-leg Anthropic usage object in Anthropic's own field order.
-fn usage_object(input: u64, cache_creation: u64, cache_read: u64, output: u64) -> Value {
-    json!({
-        "input_tokens": input,
-        "cache_creation_input_tokens": cache_creation,
-        "cache_read_input_tokens": cache_read,
-        "output_tokens": output,
-    })
-}
-
 /// Frame one named, compact, UTF-8-preserving Anthropic SSE event.
 fn event_frame(name: &str, payload: &Value) -> String {
     format!("event: {name}\ndata: {}\n\n", compact_json(payload))
@@ -984,8 +934,11 @@ impl MessagesSseEncoder {
 }
 
 mod aggregate;
+mod usage;
 
 pub use aggregate::{completed_messages_body, completed_messages_body_with_reasoning};
+pub(crate) use usage::messages_usage;
+use usage::usage_object;
 
 /// Attach the `x-experiential-ignored-parameters` disclosure to one message
 /// object when any control was dropped; an empty list adds nothing.
@@ -1006,3 +959,5 @@ pub(super) fn disclose_ignored_parameters(message: &mut Value, ignored_parameter
 mod tests;
 #[cfg(test)]
 mod tests_claude_code;
+#[cfg(test)]
+mod tests_usage;
